moved edit comp to new file
